refactoring to more closely align with main
