+++ package.json
@@ -76,7 +76,6 @@
     "jsdom": "^26.1.0",
     "@vitest/coverage-v8": "^1.6.1",
     "@vitest/ui": "^1.6.1",
-    "happy-dom": "^20.8.3",
     "prettier": "^2.8.7",
     "typescript": "^5.9.3",
     "vite": "^7.1.10",
